Add duplicates strategy for processResources task
From Gradle version 7, implicit duplicates strategies are forbidden.

This matters for us, since our processResources task apparently
creates duplicates of inspectionDescriptions/PSUnresolvedReference.html.

Ideally, we should look into why this is happening, but for now I've
added a duplicates strategy according to the documentation at

  https://docs.gradle.org/6.2.1/userguide/upgrading_version_5.html\#implicit_duplicate_strategy_for_copy_or_archive_tasks_has_been_deprecated

This is the error that we're getting in gradle 7 without this fix:

  * What went wrong:
  Execution failed for task ':processResources'.
  > Entry inspectionDescriptions/PSUnresolvedReference.html is a duplicate but no duplicate handling strategy has been set. Please refer to https://docs.gradle.org/7.0.1/dsl/org.gradle.api.tasks.Copy.html\#org.gradle.api.tasks.Copy:duplicatesStrategy for details.